k/usage: Refactor usage_manager for easier testing
- Breaking out inner looping mechanism so tests against that portion can
be made using ss::manual_clock to drive it.
